A bunch of fixes related to new Clang version and new way of patching Clang.
Fix PCH - must be built locally.
Do not reuse clang::HeaderSearch, create a fresh copy for each preprocessor.
Fix how DiagnosticOptions is created - it has to be heap allocated.
Add LLVM_OVERRIDE to PPCallbacks overloads to prevent future long debugging sessions when Clang people change signatures.
In case we decide to store some of data coming through network - convert it to from memoryview to bytes, as memory might disappear.
MacroBuilder was a bad idea. It adds a space between name and value. For us this is bad as we treat macro arguments as a part of value, not part of name.
Now that we use IncludeDirective again it is not as simple to determine whether header was included using relative path.