add missing DummyModuleLoader virtual member function implementations
This makes the code compatible with Clang 1.35 release which introduced two new
pure virtual member functions in its ModuleLoader interface.

The whole class is used only as a technicality to get the code compiling. It does
nothing and its member functions are never expected to be called so their actual
implementation is irrelevant.